pyproject.toml
[project]
name = "python-clean-code-debate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "tabulate>=0.9.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pytest>=8.4.2",
    "ruff>=0.12.7",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false
